fix: v1.6.11 — 第三/四轮审计修复：配对会话门控(#2)、cleanup快照迭代(#3)、publish失败状态对齐(#4)、去重时基monotonic(#5)、config_flow mock契约(#6)、cover浮点截断(审计4#2)；#1/#7裁决误报/维持；112测试通过

+++ huijian_mqtt_broker/custom_components/window_controller_gateway/cover.py
@@ -109,7 +109,11 @@ class WindowControllerCover(WindowControllerBaseEntity, RestoreEntity, CoverEnti
             r_travel = (device.get("attributes") or {}).get("r_travel")
             try:
                 if r_travel is not None:
-                    return int(r_travel) <= 0
+                    # v1.6.11（外部审计 #2）：int() 截断会把 0.5 判成"关"
+                    # （int(0.5)=0 ≤0），违反用户定案的">0=打开"语义。协议
+                    # 规定整数 0-100，但 JSON 可携浮点——用 float 直比，
+                    # 非数值串由既有 except (ValueError, TypeError) 落 None
+                    return float(r_travel) <= 0
             except (ValueError, TypeError):
                 pass
         return None
